Rename bookmarks to match actual page location name
